osd: implement dynamic recovery throttling
In the dynamic mode, user can set the client bw/iops threshold and the
min/max recovery bw/iops. When the client io rate on an OSD is below the
threshold, the recovery bw/iops increases in the range [min, max]. When
the client io is beyond the threshold, the recovery bw/iops drops to the
min.

Signed-off-by: Zhiqiang Wang <zhiqiang@xsky.com>