osd: avoid changing the bucket frequently in the dymamic mode
In some cases, we could see a transient big increase/decrease in
the workload. This commit tries to minimize the impaction to the
dynamic mode in this case. The general idea is that when the client
bw/iops drops below the threshold, we don't increase the bucket
right away. We do this until we see 3 consective times. This is
almost the same when the client bw/iops increases above the threshold,
with a slightly difference in counting the times.

Signed-off-by: Zhiqiang Wang <zhiqiang@xsky.com>